+++ fronts/finding/algorithms.py
@@ -1,13 +1,16 @@
 """ Front finding algorithms """
-from fronts.finding import pyboa
 import numpy as np
 
 from skimage import morphology
+
+from fronts.finding import pyboa
+from fronts.finding.sharpen import global_sharpen_pq
 
 from IPython import embed
 
 def fronts_from_gradb2(gradb2, window:int=40, thin:bool=False,
                       rm_weak:float=None, dilate:bool=False,
+                      sharpen:bool=False,
                       connectivity:int=2, threshold:float=90,
                       thresh_mode:str='generic', n_workers:int=None,
                       min_size:int=7, verbose:bool=False,
@@ -23,6 +26,8 @@ def fronts_from_gradb2(gradb2, window:int=40, thin:bool=False,
         The window size used for thresholding in the front detection algorithm.
     thin : bool, optional, default=False
         If True, thins the detected fronts to single-pixel width.
+    sharpen : bool, optional, default=False
+        If True, sharpens the detected fronts on gradb2 to single-pixel width.
     rm_weak : float, optional, default=None
         If provided, removes weak segments where gradb2 values are below this threshold.
     dilate : bool, optional, default=False
@@ -55,6 +60,11 @@ def fronts_from_gradb2(gradb2, window:int=40, thin:bool=False,
     if rm_weak is not None:
         res_frnt_np &= gradb2 > rm_weak
 
+    # Sharpen?
+    if sharpen:
+        res_frnt_np = global_sharpen_pq(res_frnt_np, gradb2,
+                                   protect_endpoints=True)
+
     # Thin?
     if thin:
         if verbose:
@@ -63,10 +73,6 @@ def fronts_from_gradb2(gradb2, window:int=40, thin:bool=False,
         res_frnt_np = morphology.thin(res_frnt_np)
         if verbose:
             print(f'There are {np.sum(res_frnt_np)} front pixels after thinning')
-        if debug:
-            print('Thinning another time...')
-            thin_2x = morphology.thin(res_frnt_np)
-            embed(header='63 of algorithms')
     
     # Crop?
     if min_size > 0:
@@ -84,7 +90,7 @@ def fronts_from_gradb2(gradb2, window:int=40, thin:bool=False,
         res_frnt_crop = morphology.dilation(res_frnt_crop)#, morphology.square(3))
 
     # Thin a final time
-    if thin:
+    if thin or sharpen:
         if verbose:
             print('Thinning a final time...')
         res_frnt_crop = morphology.thin(res_frnt_crop)
