+++ fronts/finding/configs/finding_config_D.yaml
@@ -1,0 +1,19 @@
+# Front Finding Configuration - D
+# This configuration file defines parameters for ocean front detection
+# Modifying threshold
+
+label: "D"                      # Config label - will be part of the output filename
+
+# Binary parameters
+binary:
+  # Required
+  window: 64                      # Size of the region for thresholding (pixels)
+  threshold: 80                   # Threshold for front finding
+  thresh_mode: "pool"             # Mode for finding threshold [generic, vectorized, dask, pool]
+  thin: false                     # Apply morphological thinning to fronts
+  sharpen: true                   # Apply morphological thinning to fronts
+
+  # Optional fields
+  dilate: false                   # Dilate the front (usually after thin + crop)
+  min_size: 7                     # Minimum size for front (pixels) - used for cropping
+  connectivity: 2                 # Connectivity for morphological operations (1=4-conn, 2=8-conn)
